feat: Add Day 6 Basic App
This commit adds the code for the Day 6 Basic App, which includes a widget that displays a point score and a button to open a dialog box. The dialog box allows the user to choose from a list of randomly generated points. If a point is selected, the score is updated accordingly.